work in progress
Former-commit-id: e169506b4d982889ed71c5ea1e3706b66b1f7073
Former-commit-id: 4cb8ebae98d7d91b5a870a08c4b8ef1b242b5a25

+++ src/thing_template.c
@@ -721,7 +721,7 @@ int16_t thing_template_get_health (thing_templatep t)
     return (t->health);
 }
 
-uint32_t thing_template_get_damage (thing_templatep t)
+uint16_t thing_template_get_damage (thing_templatep t)
 {
     return (t->damage);
 }
